refactor(angular/theme-service): rename preview example to match js file name
